+++ server/graphql/mutations/usergroups/removeUserGroup.js
@@ -3,6 +3,7 @@ const mongoose = require('mongoose');
 const { outputType } = require('../../types/UserGroups');
 const getProjection = require('../../get-projection');
 const emitSocketEvent = require('../../../utils/emitSocketEvent');
+const getUserPermissions = require('../../../utils/getUserPermissions');
 
 const UserGroup = mongoose.model('UserGroup');
 
@@ -16,6 +17,9 @@ module.exports = {
     },
   },
   async resolve(root, args, ctx, ast) {
+    const perms = await getUserPermissions(ctx.user._id);
+    if (!perms.users.canManageUserGroups) throw new Error('You do not have permission to manage User Groups.');
+
     const projection = getProjection(ast);
     const removedUserGroup = await UserGroup
       .findByIdAndRemove(args._id, { select: projection })
